Initial Commit
This is the initial commit for this little script which opens a spout to a given organization in LimaCharlie and examines the tags. If a given tag matches an associated colour is applied to the Phillip Hue lights on the same network. It is a rough sketch but demonstrates some of what is possible with LimaCharlie's developer friendly approach to cybersecurity tools and infrastructure.